requirements.txt
Django<=2
mysqlclient
djangorestframework
coreapi
django_cron
health-check
structlog
uuid
bagit
